+++ poetry/console/commands/debug/__init__.py
@@ -1,0 +1,1 @@
+from .resolve import DebugResolveCommand